+++ public/locales/en/common.json
@@ -434,7 +434,7 @@
         "origin_ip": "Origin IP",
         "status": "Status" 
     },
-    "pbs": {
+    "proxmoxbackupserver": {
         "datastore_usage": "Datastore",
         "failed_tasks_24h": "Failed Tasks 24h",
         "cpu_usage": "CPU",
