+++ src/utils/proxy/handlers/credentialed.js
@@ -34,7 +34,7 @@ export default async function credentialedProxyHandler(req, res, map) {
         headers.Authorization = `Bearer ${widget.key}`;
       } else if (widget.type === "proxmox") {
         headers.Authorization = `PVEAPIToken=${widget.username}=${widget.password}`;
-      } else if (widget.type === "pbs") {
+      } else if (widget.type === "proxmoxbackupserver") {
         headers = {};
         headers.Authorization = `PBSAPIToken=${widget.username}:${widget.password}`;
       } else if (widget.type === "autobrr") {
