+++ src/widgets/proxmoxbackupserver/widget.js
