+++ src/widgets/widgets.js
@@ -31,7 +31,7 @@ import ombi from "./ombi/widget";
 import opnsense from "./opnsense/widget";
 import overseerr from "./overseerr/widget";
 import paperlessngx from "./paperlessngx/widget";
-import pbs from "./pbs/widget";
+import proxmoxbackupserver from "./proxmoxbackupserver/widget";
 import pihole from "./pihole/widget";
 import plex from "./plex/widget";
 import portainer from "./portainer/widget";
@@ -93,7 +93,7 @@ const widgets = {
   opnsense,
   overseerr,
   paperlessngx,
-  pbs,
+  proxmoxbackupserver,
   pihole,
   plex,
   portainer,
